Added Polish
-updated UI
-added quality of life controls
-added shop
-fixed shop/job board bugs
-added business operations manager to shop and job board
-added tools
-added med pack to heal in field
-added escape transport to escape field from anywhere without losing items
-balanced costs
-added gene editor
-added flower varieties

+++ Assets/Scripts/GameManager/BuildAsset.cs
@@ -28,9 +28,12 @@ public static class BuildAsset
     public static int CountType(System.Type type, string[] checkFolders) =>
         AssetDatabase.FindAssets($"t:{type}", checkFolders).Length;
 
-    public static void Destroy(ScriptableObject asset)
+    public static void Destroy(ScriptableObject asset, string clonePath)
     {
         string path = AssetDatabase.GetAssetPath(asset);
-        AssetDatabase.DeleteAsset(path);
+        if(path.StartsWith(clonePath, System.StringComparison.Ordinal))
+            AssetDatabase.DeleteAsset(path);
+        else
+            Debug.Log("Canceled attempt to delete item outside of clone folder");
     }
 }
